% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amisc-package.R
\docType{package}
\name{Amisc-package}
\alias{Amisc}
\alias{Amisc-package}
\title{Amisc: Miscellaneous Utility Functions}
\description{
Provides some commonly used utility functions used
    in biostatistics, such as row and column percentages in a table, date
    formatting, survival functions, and more.
}
\author{
\strong{Maintainer}: Aline Talhouk \email{atalhouk@bccrc.ca}

Authors:
\itemize{
  \item Derek Chiu \email{dchiu@bccrc.ca}
  \item Shanzhao Wang \email{szwang@bccrc.ca}
}

}
\keyword{internal}
