% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amisc.R
\docType{package}
\name{amisc}
\alias{amisc}
\alias{amisc-package}
\title{amisc: A package with functions I find handy}
\description{
The package provides several functions I tend to use often.
}
\section{Debug functions}{


Debug functions type the result accordintg to the logical
value of variable \code{.testing}
}

