% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{chk}
\alias{chk}
\title{Helper function for debugging; prints an integer, timestamp, and optional
message}
\usage{
chk(num, msg = "")
}
\arguments{
\item{num}{A number to be printed (useful to reference a specific place in a
code file}

\item{msg}{A message to be printed}
}
\description{
A helper function for debugging that parses and outputs a
message containing a number, a timestamp, and (optionally) a message.
}
\examples{
chk(1, "code execution began")
x <- runif(10)
chk(2, "runif completed")
y <- rnorm(10)
chk(3, "rnorm completed")
}
