% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{deriv_expit}
\alias{deriv_expit}
\title{Derivative of the expit function}
\usage{
deriv_expit(x)
}
\arguments{
\item{x}{Numeric input}
}
\value{
Numeric output
}
\description{
Computes the derivative of the expit function
}
\examples{
deriv_expit(3)
}
