% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{deriv_logit}
\alias{deriv_logit}
\title{Derivative of the logit function}
\usage{
deriv_logit(x)
}
\arguments{
\item{x}{Numeric input}
}
\value{
Numeric output
}
\description{
Computes the derivative of the logit function
}
\examples{
deriv_logit(0.9)
}
