% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBy.R
\name{describeBy}
\alias{describeBy}
\title{Descriptive statistics}
\usage{
describeBy(data, var.names, var.labels = var.names, by1,
  dispersion = "se", ShowTotal = TRUE, by2 = NULL, per = "col",
  digits = 1, p.digits = 3, Missing = TRUE, stats = "parametric",
  simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{data:}{Input data.frame(df)}
}
\value{
possibly something
}
\description{
takes up variables from a data.frame(df) and returns a descriptive statistic based on a selected factor `by1` in df
}
\details{
univariable association
}
\examples{
TODO
}
\author{
Aline Talhouk
}
