% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{dis_plot}
\alias{dis_plot}
\title{Visualize one or more one-dimensional distributions}
\usage{
dis_plot(val, grp, ref_vals = NULL, common_scale = TRUE)
}
\arguments{
\item{val}{A vector of numeric data values}

\item{grp}{A vector denoting groups}

\item{ref_vals}{A named vector denoting reference values to be overlaid on
plot; see examples}

\item{common_scale}{If TRUE, a common axis scale is used for all groups}
}
\value{
A ggplot2 object
}
\description{
Produce a grid of one-dimensional jittered scatterplots.
}
\examples{
data(iris)
dis_plots(
  val = iris$Petal.Length,
  grp = iris$Species
)
dis_plots(
  val = iris$Petal.Length,
  grp = iris$Species,
  ref_vals = c(setosa=1.5, versicolor=4, virginica=5.5)
)
}
