% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{In}
\alias{In}
\title{Alias for as.integer function}
\usage{
In()
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
Integer output
}
\description{
Alias for as.integer function; see \code{as.integer}
documentation for more info.
}
\examples{
In(3!=4)
}
