% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{memoise2}
\alias{memoise2}
\title{Memoise a function}
\usage{
memoise2(fnc)
}
\arguments{
\item{fnc}{A function to be memoised}
}
\value{
A memoised version of function
}
\description{
This is a lightweight and faster version of the \code{memoise}
function from the \code{memoise} package. The speed comes at the cost of
fewer error handling checks.
}
\examples{
f <- function(x) { Sys.sleep(1); return(x^2); }
f_mem <- memoise2(f)
system.time(f_mem(4)) # First call
system.time(f_mem(4)) # Second call
}
