% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Set Package Specific Options when Package is Loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}
}
\value{
none (invisible \code{NULL})
}
\description{
This function sets package specific options.
Option names start with package name and dot (\code{amisc.*}).
}
\details{
\code{.onLoad} sets the \code{amisc.testing} option to \code{FALSE}.
This option is used to control testing typeouts in function \code{testing}.
}
\examples{
(amisc.tst <- getOption("amisc.testing"))
options(amisc.testing = TRUE)
getOption("amisc.testing")
options(amisc.testing = amisc.tst)
rm(amisc.tst)

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\link{testing}
}

