% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{pause}
\alias{pause}
\title{Pause Execution}
\usage{
pause(prompt = "Pause. Press <Enter> to continue... ")
}
\arguments{
\item{prompt}{the string printer when prompting the user to press a key.}
}
\description{
Pause stops execution and waits for any key to be pressed.
}
\examples{
\dontrun{
for (i in 1:3) {
  print(i)
  pause()
}
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

