% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{purge}
\alias{purge}
\title{Detach all Attached Datasets}
\usage{
purge(verbose = FALSE)
}
\arguments{
\item{verbose}{print final search path list for confirmaton}
}
\description{
All attached datasets will be detached. Especially useful
after a number of attaces of the same dataset
(in interactive mode it happens).
}
\examples{
bla <- data.frame(x=1,y=1)
try(print(x))
attach(bla)
attach(bla)
search()
purge()
search()
#
try(x)
attach(bla)
try(x)
purge(verbose=TRUE)
#
rm(bla)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

