% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{read.clipboard}
\alias{read.clipboard}
\title{Read Clipboard}
\usage{
read.clipboard(header = T, sep = "\\t", ...)
}
\arguments{
\item{header}{a logical value indicating whether the file contains the names of the variables as its first line. Defaults to \code{TRUE}. If missing, the value is determined from the file format: header is set to TRUE if and only if the first row contains one fewer field than the number of columns.}

\item{sep}{the field separator character (default \code{\\t}).
Values on each line of the file are separated by this character.
If sep = "" (the default for read.table) the separator
is 'white space', that is one or more spaces, tabs, newlines or carriage returns.}

\item{...}{additional arguments passed to \link{read.table}.}
}
\value{
A data frame. See \link{read.table}
}
\description{
Read text (usually table with data) from clipboard. This is a wrapper function for \code{read.table} with different defaults.
}
\examples{
# Copy some data to clipboard before running the code below
read.clipboard()
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\link{read.table}.
}

