% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{testing}
\alias{testing}
\title{Print Object Names and Values}
\usage{
testing(..., prefix = " Testing > > > \\n")
}
\arguments{
\item{...}{R objects.}
}
\value{
None (invisible NULL).
}
\description{
Outputs the objects names and values.
}
\details{
This function is useful for debugging where it can be useful to know
the name of the object associated with the displayed value.

The actual printout depends on option \code{getOption("amisc.testing")}.
If it is set to \code{FALSE} (default) nothing happens.
If set to  \code{options(amisc.testing = TRUE)} test output is produced.
}
\examples{
.oldOpt <- options(amisc.testing = TRUE)
x <- 3.14
testing(x)
y <- 1:5
testing(y, prefix="-------------> \\n...")
options(amisc.testing = FALSE)
testing(y)
options(.oldOpt) # reset option
rm(.oldOpt)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
To control options see \link{options} and \link{.onLoad}, \link{cat} for print controls.
}

