% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitestCat.R
\name{unitestsCat}
\alias{unitestsCat}
\title{some documentation}
\usage{
unitestsCat(fac.dat, fac.var, fac.label, by, per = "col", digits = 1,
  p.digits = 3, showMissing, simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{some}{parameters}
}
\value{
possibly something
}
\description{
a function to do something
}
\details{
this function will do something
}
\examples{
TODO
}
\author{
Aline Talhouk
}
