% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{Var}
\alias{Var}
\title{MLE Variance Estimator}
\usage{
Var(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical. Should missing values be remove?}
}
\value{
MLE estimate of variance. Note that it is biased.
}
\description{
Compute variance using the MLE variance estimator.
}
\examples{
Var(1:4)
3/4*var(1:4)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

