% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBy.R
\name{describeBy}
\alias{describeBy}
\title{Descriptive statistics}
\usage{
describeBy(data, var.names, var.labels = var.names, by1, by2 = NULL,
  ShowTotal = TRUE, Missing = TRUE, digits = 0, p.digits = 3,
  dispersion = c("sd", "se"), stats = c("parametric",
  "non-parametric"), per = "col", simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{data}{data.frame to produce descriptive statistics}

\item{var.names}{variable names of interest in \code{data}}

\item{var.labels}{variable descriptions. Uses \code{var.names} by default.}

\item{by1}{factor to split other variables by in \code{data}}

\item{by2}{optional second factor to split other variables by}

\item{ShowTotal}{logical; if \code{TRUE}, it shows the total number of each level
w/ \code{by1}.}

\item{Missing}{logical; if \code{TRUE}, shows missing value counts, if they exist}

\item{digits}{number of digits to round descriptive statistics}

\item{p.digits}{number of digits to round univariable test p-value}

\item{dispersion}{measure of variability, either "sd" (default) or "se".}

\item{stats}{either "parametric" (default) or "non-parametric" univariable
tests are performed}

\item{per}{print column ("col") or row ("row") percentages}

\item{simulate.p.value}{passed to \code{chisq.test}. Only relevant for categorical
variables.}

\item{B}{passed to \code{chisq.test}. Only relevant for categorical variables.}
}
\value{
A table with descriptive statistics for continuous and categorical
variables and relevant univariable association tests
}
\description{
Descriptive statistics and univariable association tests
}
\details{
Takes variables from \code{data} and returns descriptive statistics split on
factor \code{by1}.
}
\examples{
mtcars$cyl <- as.factor(mtcars$cyl)
mtcars$vs <- as.character(mtcars$vs)
Amisc::describeBy(data = mtcars, var.names = c("vs", "hp"), by1 = "cyl",
Missing = TRUE, dispersion = "sd", stats = "parametric")
}
\author{
Aline Talhouk
}
