% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBy.R
\name{describeBy}
\alias{describeBy}
\title{Descriptive statistics}
\usage{
describeBy(data, var.names, var.labels = var.names, by1,
  dispersion = "se", ShowTotal = TRUE, by2 = NULL, per = "col",
  digits = 0, p.digits = 3, Missing = TRUE, stats = "parametric",
  simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{data}{data.frame to produce descriptive statistics}

\item{var.names}{variable names of interest in `data`}

\item{var.labels}{variable descriptions. Uses `var.names` by default.}

\item{by1}{factor to split other variables by in `data`}

\item{dispersion}{measure of variability, either "se" (default) or "sd".}

\item{ShowTotal}{logical; if `TRUE`, it shows the total number of each level
w/ `by1`.}

\item{by2}{optional second factor to split other variables by}

\item{per}{print column ("col") or row ("row") percentages}

\item{digits}{number of digits to round descriptive statistics}

\item{p.digits}{number of digits to round univariable test p-value}

\item{Missing}{logical; if `TRUE`, shows missing value counts, if they exist}

\item{stats}{either "parametric" or "non-parametric" univariable tests are
performed}

\item{simulate.p.value}{passed to `chisq.test`. Only relevant for categorical
variables.}

\item{B}{passed to `chisq.test`. Only relevant for categorical variables.}
}
\value{
A table with descriptive statistics for continuous and categorical
  variables and relevant univariable association tests
}
\description{
Descriptive statistics and univariable association tests
}
\details{
Takes variables from `data` and returns descriptive statistics split on
factor `by1`.
}
\examples{
mtcars$cyl <- as.factor(mtcars$cyl)
mtcars$vs <- as.character(mtcars$vs)
Amisc::describeBy(data = mtcars, var.names = c("vs", "hp"), by1 = "cyl",
dispersion = "sd", Missing = TRUE, stats = "parametric")
}
\author{
Aline Talhouk
}
