% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBy.R
\name{describeBy}
\alias{describeBy}
\title{Descriptive statistics}
\usage{
describeBy(
  data,
  var.names,
  var.labels = var.names,
  by1,
  by2 = NULL,
  total = c("top", "bottom", "none"),
  Missing = TRUE,
  test = TRUE,
  digits = 0,
  p.digits = 3,
  dispersion = c("sd", "se"),
  stats = c("parametric", "non-parametric"),
  per = "col",
  simulate.p.value = FALSE,
  B = 2000,
  bold_var = TRUE,
  fill_var = FALSE
)
}
\arguments{
\item{data}{data.frame to produce descriptive statistics}

\item{var.names}{variable names of interest in \code{data}}

\item{var.labels}{variable descriptions. Uses \code{var.names} by default.}

\item{by1}{factor to split other variables by in \code{data}}

\item{by2}{optional second factor to split other variables by}

\item{total}{add a row showing the total counts of each \code{by1} level at the
\code{top} or \code{bottom} of the table. Setting \code{none} hides the total row.}

\item{Missing}{logical; if \code{TRUE}, shows missing value counts, if they exist}

\item{test}{logical; if \code{TRUE}, univariable tests are performed and a
\code{PValue} column is added to the end of the table.}

\item{digits}{number of digits to round descriptive statistics}

\item{p.digits}{number of digits to round univariable test p-value}

\item{dispersion}{measure of variability, either "sd" (default) or "se".}

\item{stats}{either "parametric" (default) or "non-parametric" univariable
tests are performed for continuous variables. We use the parametric one-way
test or the non-parametric Kruskal-Wallis test.}

\item{per}{print column ("col") or row ("row") percentages}

\item{simulate.p.value}{passed to \code{chisq.test}. Only relevant for categorical
variables.}

\item{B}{passed to \code{chisq.test}. Only relevant for categorical variables.}

\item{bold_var}{logical; if \code{TRUE}, the \code{Variable} names are wrapped in
double asterisks. If the table is parsed by pandoc the variable names are
in bold.}

\item{fill_var}{logical; if \code{TRUE}, the \code{Variable} name is repeated for every
row it pertains to. If \code{FALSE}, the name is only shown when it changes.}
}
\value{
A table with descriptive statistics for continuous and categorical
variables and relevant univariable association tests
}
\description{
Descriptive statistics and univariable association tests
}
\details{
Takes variables from \code{data} and returns descriptive statistics split on
factor \code{by1}.
}
\examples{
mtcars$cyl <- as.factor(mtcars$cyl)
mtcars$vs <- as.character(mtcars$vs)
Amisc::describeBy(data = mtcars, var.names = c("vs", "hp"), by1 = "cyl",
Missing = TRUE, dispersion = "sd", stats = "parametric")
}
\author{
Aline Talhouk
}
